requirements.txt
mlflow
torch
scikit-learn
numpy
pandas
sentence-transformers
faiss-cpu
openai
tqdm